pid: eyeballed for a fully charged battery
Made a video showing the result. Apparently the center of mass is not exactly at the center of the
robot and the battery needs to stick out a bit at the front to rebalance it.

Also my desk is a bit tilted so the pitch is heavily affected by the yaw, the video shows it
clearly. Next step: a inner control loop for the yaw.